shelf_detector/requirements.txt
pyzbar
Pillow
qrcode
pyyaml